chore: refactor of structures
