+++ src/drug_discovery/docking.py
@@ -13,6 +13,7 @@ from deeporigin.exceptions import DeepOriginException
 from deeporigin.tools.utils import get_statuses_and_progress, query_run_statuses
 from deeporigin.utils.core import PrettyDict, hash_strings
 from deeporigin_molstar import DockingViewer, JupyterViewer
+from deeporigin.drug_discovery.structures import Ligand, ligands_to_dataframe
 
 Number = float | int
 
@@ -34,7 +35,7 @@ class Docking:
 
         df1 = self.parent.get_csv_results_for("Docking")
 
-        df2 = chem.ligands_to_dataframe(self.parent.ligands)
+        df2 = ligands_to_dataframe(self.parent.ligands)
         df2["SMILES"] = df2["Ligand"]
         df2.drop(columns=["Ligand"], inplace=True)
