+++ src/functions/docking.py
@@ -1,17 +1,31 @@
+"""This module implements a low level function to perform molecular docking using the Deep Origin API.
+
+The main function `dock()` takes a Protein object, a list of ligand SMILES strings, and docking box parameters
+to perform docking calculations. The docking box can be specified either by providing explicit coordinates for the
+pocket center, or by passing a Pocket object which contains the pocket center information.
+
+The module interfaces with the Deep Origin docking service to perform the actual docking calculations remotely.
+"""
+
 import base64
+import hashlib
+import json
+import os
 import requests
+from pathlib import Path
 from beartype import beartype
 from typing import Tuple, Optional
 from deeporigin.drug_discovery.structures import Pocket, Protein
 from deeporigin.exceptions import DeepOriginException
 
 URL = "http://docking.default.jobs.edge.deeporigin.io/dock"
+CACHE_DIR = os.path.expanduser("~/.deeporigin/docking")
 
 
 @beartype
 def dock(
     protein: Protein,
-    smiles_list: list[str],
+    smiles_string: str,
     box_size: Tuple[float, float, float] = (20.0, 20.0, 20.0),
     pocket_center: Optional[Tuple[int, int, int]] = None,
     pocket: Optional[Pocket] = None,
@@ -35,40 +49,65 @@ def dock(
     if pocket_center is None:
         raise DeepOriginException("Pocket center is required")
 
-    # Read and encode the protein file
+    # Create hash of inputs
+    hasher = hashlib.sha256()
+
+    # Hash protein file contents
     with open(protein.file, "rb") as f:
-        encoded_protein = base64.b64encode(f.read()).decode("utf-8")
+        hasher.update(f.read())
 
-    # Prepare the request payload
-    payload = {
-        "functionId": "docking",
-        "params": {
-            "protein": encoded_protein,
-            "smiles_list": smiles_list,
-            "box_size": list(box_size),
-            "pocket_center": list(pocket_center),
-        },
-    }
+    # Hash pocket file if provided
+    if pocket is not None and pocket.file is not None:
+        with open(pocket.file, "rb") as f:
+            hasher.update(f.read())
 
-    # Make the API request
-    response = requests.post(
-        URL,
-        json=payload["params"],
-        headers={"Content-Type": "application/json"},
+    # Hash other inputs
+    hasher.update(
+        json.dumps(
+            {
+                "smiles": smiles_string,
+                "box_size": list(box_size),
+                "pocket_center": list(pocket_center),
+            }
+        ).encode()
     )
 
-    # Raise an exception for bad status codes
-    response.raise_for_status()
+    cache_hash = hasher.hexdigest()
+    sdf_file = str(Path(CACHE_DIR) / f"{cache_hash}.sdf")
 
-    response = response.json()
+    # Check if cached result exists
+    if not os.path.exists(sdf_file):
+        # Read and encode the protein file
+        with open(protein.file, "rb") as f:
+            encoded_protein = base64.b64encode(f.read()).decode("utf-8")
 
-    # return response
+        # Prepare the request payload
+        payload = {
+            "functionId": "docking",
+            "params": {
+                "protein": encoded_protein,
+                "smiles_list": [smiles_string],
+                "box_size": list(box_size),
+                "pocket_center": list(pocket_center),
+            },
+        }
 
-    sdf_file = "temp.sdf"
+        # Make the API request
+        response = requests.post(
+            URL,
+            json=payload["params"],
+            headers={"Content-Type": "application/json"},
+        )
 
-    with open(sdf_file, "w") as file:
-        for solution in response[0]["solutions"]:
-            file.write(solution["output_sdf_content"])
+        # Raise an exception for bad status codes
+        response.raise_for_status()
+
+        response = response.json()
+
+        # Write SDF file to cache
+        with open(sdf_file, "w") as file:
+            for solution in response[0]["solutions"]:
+                file.write(solution["output_sdf_content"])
 
     from deeporigin_molstar import DockingViewer, JupyterViewer
